Move Readtags impl to dumb_analyzer
