+++ crates/maple_cli/src/dumb_analyzer/find_usages/search_engine/ctags.rs
@@ -1,1 +1,184 @@
-fn hello() {}
+use std::hash::Hash;
+use std::ops::Deref;
+use std::path::{Path, PathBuf};
+use std::str::FromStr;
+
+use anyhow::Result;
+use filter::subprocess::Exec;
+use itertools::Itertools;
+
+use crate::paths::AbsPathBuf;
+use crate::tools::ctags::{TagsConfig, EXCLUDE, TAGS_DIR};
+
+enum FilteringType {
+    StartWith,
+    Contain,
+    Inherit,
+}
+
+/// Represents the manager of tags file.
+#[derive(Debug, Clone, Eq, PartialEq, Hash)]
+pub struct Readtags<'a, P> {
+    config: TagsConfig<'a, P>,
+    tags_path: PathBuf,
+}
+
+impl<'a, P: AsRef<Path> + Hash> Readtags<'a, P> {
+    pub fn new(config: TagsConfig<'a, P>) -> Self {
+        let tags_path = config.tags_path();
+        Self { config, tags_path }
+    }
+
+    /// Returns `true` if the tags file already exists.
+    pub fn exists(&self) -> bool {
+        self.tags_path.exists()
+    }
+
+    pub fn create(&self) -> Result<()> {
+        self.config.generate_tags()
+    }
+
+    fn build_exec(&self, query: &str, filtering_type: FilteringType) -> Exec {
+        // https://docs.ctags.io/en/latest/man/readtags.1.html#examples
+        let cmd = Exec::cmd("readtags")
+            .arg("--tag-file")
+            .arg(&self.tags_path)
+            .arg("-E")
+            .arg("-ne");
+
+        let cmd = if query.chars().all(char::is_lowercase) {
+            cmd.arg("--icase-match")
+        } else {
+            cmd
+        };
+
+        match filtering_type {
+            FilteringType::StartWith => cmd.arg("--prefix-match").arg("-").arg(query),
+            FilteringType::Contain => cmd
+                .arg("-Q")
+                .arg(format!("(substr? (downcase $name) \"{}\")", query))
+                .arg("-l"),
+            FilteringType::Inherit => {
+                todo!("Inherit")
+            }
+        }
+    }
+
+    pub fn search(
+        &self,
+        query: &str,
+        force_generate: bool,
+    ) -> Result<impl Iterator<Item = TagLine>> {
+        use std::io::BufRead;
+
+        if force_generate || !self.exists() {
+            self.create()?;
+        }
+
+        let stdout = self
+            .build_exec(query, FilteringType::StartWith)
+            .stream_stdout()?;
+
+        Ok(std::io::BufReader::new(stdout)
+            .lines()
+            .flatten()
+            .filter_map(|line| line.parse::<TagLine>().ok()))
+    }
+}
+
+#[derive(Default, Debug)]
+pub struct TagLine {
+    pub name: String,
+    pub path: String,
+    pub pattern: String,
+    pub kind: String,
+    pub language: String,
+    pub scope: Option<String>,
+    pub line: u64,
+}
+
+impl FromStr for TagLine {
+    type Err = ();
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let mut items = s.split('\t');
+
+        let mut l = TagLine {
+            name: items.next().ok_or(())?.into(),
+            path: items.next().ok_or(())?.into(),
+            ..Default::default()
+        };
+
+        // https://docs.ctags.io/en/latest/man/ctags-client-tools.7.html#parse-readtags-output
+        if let Some(p) = items
+            .clone()
+            .peekable()
+            .peek()
+            .and_then(|p| p.strip_suffix(";\""))
+        {
+            let search_pattern_used = (p.starts_with('/') && p.ends_with('/'))
+                || (p.len() > 1 && p.starts_with('$') && p.ends_with('$'));
+            if search_pattern_used {
+                let pat = items.next().ok_or(())?;
+                let pat_len = pat.len();
+                if p.starts_with("/^") || p.starts_with("?^") {
+                    if p.ends_with("$/") || p.ends_with("$?") {
+                        l.pattern = String::from(&pat[2..pat_len - 4]);
+                    } else {
+                        l.pattern = String::from(&pat[2..pat_len - 2]);
+                    }
+                } else {
+                    l.pattern = String::from(&pat[2..pat_len]);
+                }
+            } else {
+                return Err(());
+            }
+        } else {
+            return Err(());
+        }
+
+        for item in items {
+            if let Some((k, v)) = item.split_once(':') {
+                if v.is_empty() {
+                    continue;
+                }
+                match k {
+                    "kind" => l.kind = v.into(),
+                    "language" => l.language = v.into(),
+                    "roles" | "access" => {}
+                    "scope" => l.scope = Some(v.into()),
+                    "line" => l.line = v.parse().expect("line is an integer"),
+                    unknown => {
+                        tracing::debug!(line = %s, "Unknown field: {}", unknown);
+                    }
+                }
+            }
+        }
+
+        Ok(l)
+    }
+}
+
+impl TagLine {
+    pub fn grep_format(&self, query: &str, ignorecase: bool) -> (String, Option<Vec<usize>>) {
+        let mut formatted = format!("[ctags]{}:{}:1:", self.path, self.line);
+
+        let found = if ignorecase {
+            self.pattern.to_lowercase().find(&query.to_lowercase())
+        } else {
+            self.pattern.find(query)
+        };
+
+        let indices = if let Some(idx) = found {
+            let start = formatted.len() + idx;
+            let end = start + query.len();
+            Some((start..end).into_iter().collect())
+        } else {
+            None
+        };
+
+        formatted.push_str(&self.pattern);
+
+        (formatted, indices)
+    }
+}
