+++ crates/maple_cli/src/dumb_analyzer/mod.rs
@@ -1,3 +1,3 @@
 mod find_usages;
 
-pub use self::find_usages::{RegexSearcher, UsagesInfo};
+pub use self::find_usages::{Readtags, RegexSearcher, UsagesInfo};
